Add doc and split file
